requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
mlflow==2.7.1
streamlit==1.27.0
dvc==3.25.0
pillow==10.0.1
opencv-python-headless==4.8.1.78
plotly==5.17.0
nbformat==5.9.2
ipykernel==6.25.2
